Conditionally enable workflows
Only run metal, specular and diffuse if such material is present.
No noticeable effect could be measured in benchmark.
Results appear to be quite noisy, though.
Meaning, sometimes significant improvements could be measured, but not the next time.
This indicates that this is probably a useful technique, but some issue still persists.